Logging enabled. Refactored API operation.

+++ src/main/java/com/ayan/travel/hotel/api/service/HotelAddressService.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,7 +23,6 @@ import com.ayan.travel.hotel.service.DateTimeService;
 public class HotelAddressService {
 
 	private static final String DATE_FORMATTER = "dd-MM-yyyy'T'HH:mm:ss'Z'";
-	private static final String HOTEL_ID_MISSING = "A mandatory field 'Hotel Id' is missing";
 	private static final String HOTEL_ADDRESS_TYPE_MISSING = "A mandatory field 'Hotel address type' is missing";
 	private static final String HOTEL_ADDRESS_LINE_1_MISSING = "A mandatory field 'Hotel address line 1' is missing";
 	private static final String HOTEL_CITY_MISSING = "A mandatory field 'Hotel address city' is missing";
@@ -32,7 +33,10 @@ public class HotelAddressService {
 	private static final String RESORCE_ALREADY_UPDATED = "The requested resource has already been updated by another user since it is been read";
 	private static final String HOTEL_ADDRESS_ALREADY_EXISTS = "A hotel address with requested hotel code and address type exists in the system";
 	private static final String HOTEL_ADDRESS_NOT_EXISTS = "A hotel address with requested hotel code and address type doesn't exist in the system";
-
+	private static final String HOTEL_NOT_EXISTS = "A hotel with requested hotel code doesn't exists in the system";
+	
+	private static final Logger LOGGER = LoggerFactory.getLogger(HotelAddressService.class);
+	
 	private HotelAddressRepository hotelAddressRepository;
 	private HotelService hotelService;
 
@@ -42,14 +46,20 @@ public class HotelAddressService {
 		this.hotelService = hotelService;
 	}
 
-	public HotelAddress createHotelAddress(HotelAddressRequestDTO input, Long hotelId) {
+	public HotelAddress createHotelAddress(HotelAddressRequestDTO input, String hotelCode) {
 		DateTimeService dateTimeService = new DateTimeService();
 		String createdAt = dateTimeService.getDateTime(DATE_FORMATTER);
-		validateMandatoryFields(input, hotelId);
 		
+		if (LOGGER.isInfoEnabled()) {
+			LOGGER.info("Hotel address create operation started at {}", createdAt);
+		}
+		
+		validateMandatoryFields(input, hotelCode);
+		
+		Long hotelId = hotelService.getHotelByCode(hotelCode).getHotelId();
 		Hotel hotel = hotelService.getHotelById(hotelId);
 		
-		if(getAddressByType(hotelId, input.getAddressType()) != null) {
+		if(getAddressByType(hotelCode, input.getAddressType()) != null) {
 			throw new ElementExistsException(HOTEL_ADDRESS_ALREADY_EXISTS);
 			
 		} else {
@@ -60,21 +70,33 @@ public class HotelAddressService {
 		}
 	}
 
-	public HotelAddress updateHotelAddress(HotelAddressRequestDTO input, Long hotelId) {
+	public HotelAddress updateHotelAddress(HotelAddressRequestDTO input, String hotelCode) {
 		DateTimeService dateTimeService = new DateTimeService();
 		String updatedAt = dateTimeService.getDateTime(DATE_FORMATTER);
-		validateMandatoryFields(input, hotelId);
 		
-		if(StringUtils.isBlank(input.getModifiedAt())) {
-			throw new ElementUpdatedException(RESORCE_ALREADY_UPDATED);
-			
-		} else if(getAddressByType(hotelId, input.getAddressType()) == null) {
+		if (LOGGER.isInfoEnabled()) {
+			LOGGER.info("Hotel address update operation started at {}", updatedAt);
+		}
+		
+		validateMandatoryFields(input, hotelCode);
+		
+		HotelAddress hotelAddress = getAddressByType(hotelCode, input.getAddressType());
+		
+		if(hotelAddress == null) {
 			throw new NoSuchElementException(HOTEL_ADDRESS_NOT_EXISTS);
 			
 		} else {
-			HotelAddress hotelAddress = getAddressByType(hotelId, input.getAddressType());
 			
-			if(StringUtils.isBlank(hotelAddress.getModifiedAt())) {
+			if (LOGGER.isInfoEnabled()) {
+				LOGGER.info("Latest Update Date time {}", hotelAddress.getModifiedAt());
+			}
+			
+			if(StringUtils.isBlank(input.getModifiedAt()) && 
+					StringUtils.isNotBlank(hotelAddress.getModifiedAt())) {
+				throw new ElementUpdatedException(RESORCE_ALREADY_UPDATED);
+				
+			} else if (StringUtils.isBlank(hotelAddress.getModifiedAt())
+					|| input.getModifiedAt().equals(hotelAddress.getModifiedAt())) {
 				hotelAddress.setAddressLine1(input.getAddressLine1());
 				hotelAddress.setAddressLine2(input.getAddressLine2());
 				hotelAddress.setAddressLine3(input.getAddressLine3());
@@ -93,20 +115,49 @@ public class HotelAddressService {
 		}
 	}
 	
-	public void deleteAllHotelAddresses(Long hotelId) {
-		Hotel hotel = hotelService.getHotelById(hotelId);
+	public void deleteAllHotelAddresses(String hotelCode) {
+		
+		DateTimeService dateTimeService = new DateTimeService();
+		String deletedAt = dateTimeService.getDateTime(DATE_FORMATTER);
+
+		if (LOGGER.isInfoEnabled()) {
+			LOGGER.info("Hotel delete operation started at {}", deletedAt);
+		}
+		
+		if (hotelService.getHotelByCode(hotelCode) == null) {
+			throw new NoSuchElementException(HOTEL_NOT_EXISTS);
+		}
+		
+		Hotel hotel = hotelService.getHotelByCode(hotelCode);
 		List<HotelAddress> hotelAddresses = hotelAddressRepository.findByHotel(hotel);
 		hotelAddressRepository.deleteAll(hotelAddresses);
 		
 	}
 	
-	public void deleteHotelAddress(Long hotelId, String addressType) {
-		HotelAddress hotelAddress = getAddressByType(hotelId, addressType);
+	public void deleteHotelAddress(String hotelCode, String addressType) {
+		
+		DateTimeService dateTimeService = new DateTimeService();
+		String deletedAt = dateTimeService.getDateTime(DATE_FORMATTER);
+
+		if (LOGGER.isInfoEnabled()) {
+			LOGGER.info("Hotel delete operation started at {}", deletedAt);
+		}
+		
+		if (hotelService.getHotelByCode(hotelCode) == null) {
+			throw new NoSuchElementException(HOTEL_NOT_EXISTS);
+		}
+		
+		HotelAddress hotelAddress = getAddressByType(hotelCode, addressType);
 		hotelAddressRepository.delete(hotelAddress);
 	}
 
-	public HotelAddress getAddressByType(Long hotelId, String addressTypeStr) {
-		Hotel hotel = hotelService.getHotelById(hotelId);
+	public HotelAddress getAddressByType(String hotelCode, String addressTypeStr) {
+		
+		if (hotelService.getHotelByCode(hotelCode) == null) {
+			throw new NoSuchElementException(HOTEL_NOT_EXISTS);
+		}
+		
+		Hotel hotel = hotelService.getHotelById(hotelService.getHotelByCode(hotelCode).getHotelId());
 		AddressType addressType = AddressType.findByLabel(addressTypeStr);
 		
 		if (hotelAddressRepository.findByHotelAndAddressType(hotel, addressType).size() == 0) {
@@ -114,18 +165,27 @@ public class HotelAddressService {
 		} else if (hotelAddressRepository.findByHotelAndAddressType(hotel, addressType).size() > 1) {
 			throw new RuntimeException("More than one hotel Address exists for the adress type " + addressTypeStr);
 		} else {
-			return hotelAddressRepository.findByHotelAndAddressType(hotel, addressType).get(0);
+			HotelAddress hotelAddress = hotelAddressRepository.findByHotelAndAddressType(hotel, addressType).get(0);
+			if (LOGGER.isInfoEnabled()) {
+				LOGGER.info("Hotel address details {}", hotelAddress);
+			}
+			return hotelAddress;
 		}
 	}
 	
-	public List<HotelAddress> getAllAddresses(Long hotelId) {
-		return hotelAddressRepository.findByHotel(hotelService.getHotelById(hotelId));
+	public List<HotelAddress> getAllAddresses(String hotelCode) {
+		
+		if (hotelService.getHotelByCode(hotelCode) == null) {
+			throw new NoSuchElementException(HOTEL_NOT_EXISTS);
+		}
+		
+		return hotelAddressRepository.findByHotel(hotelService.getHotelByCode(hotelCode));
 		
 	}
 	
-	private Boolean validateMandatoryFields(HotelAddressRequestDTO input, Long hotelId) {
-		if (hotelId == null) {
-			throw new InputMappingException(HOTEL_ID_MISSING);
+	private Boolean validateMandatoryFields(HotelAddressRequestDTO input, String hotelCode) {
+		if (hotelService.getHotelByCode(hotelCode) == null) {
+			throw new NoSuchElementException(HOTEL_NOT_EXISTS);
 
 		} else if (StringUtils.isBlank(input.getAddressType())) {
 			throw new InputMappingException(HOTEL_ADDRESS_TYPE_MISSING);
